+++ CHANGELOG.md
@@ -1,3 +1,69 @@
+# [1.5.0](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.4.0...v1.5.0) (2025-03-26)
+
+
+### Features
+
+* actualiza el yaml del consumer de images ([2a53726](https://github.com/adriantoral/ingenieria-datos-viabi/commit/2a53726806eaae1660799baa59d01698f06ed976))
+* actualiza el yaml del consumer de metadata ([de4f243](https://github.com/adriantoral/ingenieria-datos-viabi/commit/de4f243d080b70153730d0d452be594cb7ee5534))
+* divide consumer en dos partes minio e influx ([9e62934](https://github.com/adriantoral/ingenieria-datos-viabi/commit/9e62934d72843504eecb44edf83befdfe16f1040))
+
+# [1.4.0](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.3.0...v1.4.0) (2025-03-26)
+
+
+### Features
+
+* separa el envio de metadatos y video en dos colas distintas ([c89bd51](https://github.com/adriantoral/ingenieria-datos-viabi/commit/c89bd51b7d1cf07290231e3bd9354d33c2a7ed27))
+
+# [1.3.0](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.2.4...v1.3.0) (2025-03-10)
+
+
+### Features
+
+* agrega producer en local sin kubernetes ([9404c91](https://github.com/adriantoral/ingenieria-datos-viabi/commit/9404c91b0490dd65013850d169975840198a8233))
+
+## [1.2.4](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.2.3...v1.2.4) (2025-03-05)
+
+
+### Bug Fixes
+
+* actualiza el consumer para enviar ack manual ([a8db6e9](https://github.com/adriantoral/ingenieria-datos-viabi/commit/a8db6e926f586ea1ba3e25484356b9090c7552fc))
+
+## [1.2.3](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.2.2...v1.2.3) (2025-03-05)
+
+
+### Bug Fixes
+
+* actualiza las credenciales de acceso ([ad94932](https://github.com/adriantoral/ingenieria-datos-viabi/commit/ad94932d035e94baef203146e9a0f1fc5abfda40))
+
+## [1.2.2](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.2.1...v1.2.2) (2025-02-26)
+
+
+### Bug Fixes
+
+* elimina los servicios obsoletos ([ddb8831](https://github.com/adriantoral/ingenieria-datos-viabi/commit/ddb88311dbb3db41a4e5c80cb0cb4ac26f4ed199))
+
+## [1.2.1](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.2.0...v1.2.1) (2025-02-26)
+
+
+### Bug Fixes
+
+* actualiza los deployments de influxdb, minio, rabbitmq y el consumer ([12dfe7f](https://github.com/adriantoral/ingenieria-datos-viabi/commit/12dfe7ff659718432cd1c38fae0103360a0119b9))
+
+# [1.2.0](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.1.0...v1.2.0) (2025-02-26)
+
+
+### Features
+
+* add producer and cosumer ([a5c80ba](https://github.com/adriantoral/ingenieria-datos-viabi/commit/a5c80bac8abbc1ff7616595fba6021e31747c445))
+
+# [1.1.0](https://github.com/adriantoral/ingenieria-datos-viabi/compare/v1.0.0...v1.1.0) (2025-02-26)
+
+
+### Features
+
+* add yaml for influxdb and minio deployments ([316a2cd](https://github.com/adriantoral/ingenieria-datos-viabi/commit/316a2cd40330bd572fda5550f8004b1b1fdf2ae4))
+* add yaml for influxdb and minio deployments ([66f98db](https://github.com/adriantoral/ingenieria-datos-viabi/commit/66f98db3bae60c18f1600ee9b0d1b66d706ad6a7))
+
 # 1.0.0 (2025-02-26)